+++ flying_state_machines/__init__.py
@@ -1,2 +1,3 @@
+from .asynclasses import AsyncFSM, AsyncTransition
 from .classes import Transition, FSM
 from .version import __version__, version